requirements.txt
# Aegis Threat Detection System - Dependencies

# Core ML/Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
imblearn>=0.0

# Deep Learning (optional but recommended for GPS/Login models)
tensorflow>=2.13.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AWS Integration
boto3>=1.29.0

# Dashboard
streamlit>=1.28.0

# Additional utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
